// src/constants/enums.ts
export const enum UILabel {
  Username = "Username",
  Password = "Password",
  Login_Button = "Login",
  Dashboard = "Dashboard",
  UserRole = "User Role",
  UserRoleOption = "User Role Option",
  Admin = "Admin",
  Ladies_Outwear = "Ladies Outwear",
  Product = "Product",
  Quantity = "Quantity",
  AddToCart = "Add  ToCart",
  ViewCart = "View Cart",
  CartQuantity = "Cart Quantity",
  Name = "Name",
  Email = "Email",
  Phone = "Phone",
  TellAbout = "Tell About",
  SubmitButton = "Submit Button",
  SingleClickButton = "Single Click Button",
  singleClickMsg = "Single Click Message",
  DoubleClickButton = "Double Click Button",
  doubleClickMsg = "Double Click Message",
  RightClickButton = "Right Click Button",
  RightClickMsg = "Right Click Message",
  CheckBox_A = "CheckBox A",
  RadioOne = "Radio One",
  Contact_Link = "Contact Link",
  Submit = "Submit",
  Forename = "Forename",
  Message = "Message",
  StartShopping = "StartShopping",
  Cart = "Cart",
  Checkout = "Checkout",
}

export const enum FieldType {
  Input = "input",
  Button = "button",
  Select = "select",
  Dropdown = "dropdown",
  Checkbox = "checkbox",
  Radio = "radio",
  Textarea = "textarea",
  RichTextArea = "rich-text-area",
}

export const enum RecordType {
  BasicForm = "BasicForm",
  Create = "Create",
  Update = "Update",
  Action = "Action",
}